q5/requirements.txt
langchain
langchain-core
langchain-community
google-generativeai
